Use uintptr_t for pointer int cast in SMBC_getdents_ctx()
On i386, unsigned long long is 64 bit while the pointer is 32, and
this fails under autobuild with -WError.


Autobuild-User(master): Jeremy Allison <jra@samba.org>
Autobuild-Date(master): Tue Jul 21 05:14:20 CEST 2015 on sn-devel-104

+++ source3/libsmb/libsmb_dir.c
@@ -1129,7 +1129,7 @@ SMBC_getdents_ctx(SMBCCTX *context,
 
 		/* Try and align the struct for the next entry
 		   on a valid pointer boundary by appending zeros */
-		while((rem > 0) && ((unsigned long long)ndir & (sizeof(void*) - 1))) {
+		while((rem > 0) && ((uintptr_t)ndir & (sizeof(void*) - 1))) {
 			*ndir = '\0';
 			rem--;
 			ndir++;
